model repository load refactorying
