--- back/proxy/nginx.conf
@@ -1,64 +1,0 @@
-user nginx;
-worker_processes 1;
-
-error_log /var/log/nginx/error.log warn;
-pid /var/run/nginx.pid;
-
-events {
-    worker_connections 2048;
-    multi_accept on;
-    use epoll;
-}
-
-http {
-    keepalive_timeout 10;
-    server_tokens off;
-    include /etc/nginx/mime.types;
-    default_type application/octet-stream;
-
-    gzip on;
-    gzip_disable "msie6";
-
-    gzip_comp_level 6;
-    gzip_min_length 100;
-    gzip_buffers 16 8k;
-    gzip_proxied any;
-
-    gzip_types
-    text/plain
-    application/json
-    application/xml
-    application/rss+xml
-    image/svg+xml;
-
-    reset_timedout_connection on;
-
-    upstream docker-express {
-        server express:3000;
-    }
-
-    server {
-        listen 80;
-        server_name api.sunrinsecurity.com;
-        add_header 'Access-Control-Allow-Origin' '*';
-        proxy_set_header X-Forwarded-For $remote_addr;
-        proxy_set_header X-Forwarded-Proto $scheme;
-        proxy_set_header X-Forwarded-Host $host;
-        proxy_set_header Host $http_host;
-        client_body_timeout 5s;
-        client_header_timeout 5s;
-        location / {
-            proxy_http_version 1.1;
-            proxy_pass http://docker-express;
-        }
-
-    }
-    log_format main '$remote_addr - $remote_user [$time_local] "$request" '
-    '$status $body_bytes_sent "$http_referer" '
-    '"$http_user_agent" "$http_x_forwarded_for"';
-
-    access_log /var/log/nginx/access.log main;
-
-    sendfile on;
-    include /etc/nginx/conf.d/*.conf;
-}
